Add files via upload
A project on computing the autocorrelation function and getting the best linear predictor to solve aproximate a squence of numbers that are missing for a key. This project focusees on understanding the best liinear predictor in the field of financial time series by creating functions that reflect the general form of the gamma matrix and vector as well as the acf and pacf.