Add files via upload
This time series project focuses on analizing data and properly fitiing the models. We mainly use ARMA-GARCH models for fitting and look to find the best fit that minimizes the information criteria. Additionaly we do a Value at Risk analisis that to see the performance. The data used is the log(returns) of the S&P 500 from 2007-2015.